Move Redirect class to Helpers namespace
